test(storage): fold the latest-revision compile tests into compiler_test.go
The tested function is compileList, which lives in compiler.go, so the
assertions belong beside the other compile tests rather than in a file
named after the statement that supplies the conjunct.
